move match_template to imgops
